requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
Flask~=2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.1
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pymongo~=4.2.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug~=2.2.2
WTForms==3.0.1
zipp==3.8.0

python-dotenv~=0.21.0